+++ internal/vision/visionproxy/service_test.go
