+++ internal/vision/visionproxy/stub.go
