+++ internal/vision/visionproxy/vision_proxy_client.go
