+++ internal/vision/visionproxy/vision_proxy_e2e_test.go
