+++ internal/vision/visionproxy/vision_proxy_regression_test.go
