+++ internal/vision/visionproxy/vision_proxy_test.go
